Make the heading of definitions more obvious
Definitions are generally structured as such:

Word (some info about the word)
1. Definition A
2. Definition B

However, words in other languages (such as Romanian) don't have the same entry standard, as they all seem to start directly at the list items, ignoring the "Word (some info about the word)" part.

It is important to mark headings as such (by adding a # at to the beginning of their text) because the indexes of examples don't take them into account (so the the example at index 0 is linked to the first list element - "1. Definition A", in this case).

But since such headings don't appear in all languages, you wouldn't know whether to ignore or not the first element in the definitions list, unless its nature was stated explicitly.

It also helps with correctly rendering the text in a website - so you'd know to make it a heading and not a list item (and vice versa).